feat(signals): introduce pluggable signals provider abstraction (default_v1)
